♻️[Refactor] 지도 선택 컴포넌트 구조 개선 및 지도 연결 정보 강화
- InlineMapPickerState 및 SelectedPlaceAnnotation을 개별 파일로 분리하여 모듈화
- Apple Maps 앱 열기 시 장소 이름(locationName)이 표시되도록 개선
- 파일 헤더의 작성자 정보(Author) 및 생성 정보 정리
