+++ AppProduct/AppProduct/Features/Notice/Data/NoticeReadRecord.swift
@@ -2,7 +2,7 @@
 //  NoticeReadRecord.swift
 //  AppProduct
 //
-//  Created by OpenAI on 3/12/26.
+//  Created by euijjang97 on 3/12/26.
 //
 
 import Foundation
